+++ pandas/tests/indexes/multi/test_missing.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import pandas.util.testing as tm
 import pytest
-from pandas import Int64Index, MultiIndex, PeriodIndex, UInt64Index, isna
+from pandas import Int64Index, MultiIndex, PeriodIndex, UInt64Index
 from pandas._libs.tslib import iNaT
 from pandas.core.indexes.datetimelike import DatetimeIndexOpsMixin
 
@@ -78,27 +78,9 @@ def test_nulls(idx):
     # this is really a smoke test for the methods
     # as these are adequately tested for function elsewhere
 
-    # TODO: Remove or Refactor. MultiIndex not Implemeted.
-    for name, index in [('idx', idx), ]:
-        if len(index) == 0:
-            tm.assert_numpy_array_equal(
-                index.isna(), np.array([], dtype=bool))
-        elif isinstance(index, MultiIndex):
-            idx = index.copy()
-            msg = "isna is not defined for MultiIndex"
-            with tm.assert_raises_regex(NotImplementedError, msg):
-                idx.isna()
-        else:
-
-            if not index.hasnans:
-                tm.assert_numpy_array_equal(
-                    index.isna(), np.zeros(len(index), dtype=bool))
-                tm.assert_numpy_array_equal(
-                    index.notna(), np.ones(len(index), dtype=bool))
-            else:
-                result = isna(index)
-                tm.assert_numpy_array_equal(index.isna(), result)
-                tm.assert_numpy_array_equal(index.notna(), ~result)
+    msg = "isna is not defined for MultiIndex"
+    with tm.assert_raises_regex(NotImplementedError, msg):
+        idx.isna()
 
 
 @pytest.mark.xfail
